2019 - day 19

+++ 2019/day19/test.h
@@ -4,5 +4,5 @@
 
 namespace aoc2019_day19 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    unsigned long long part_2(std::string_view path);
 }
